requirements.txt
tk
customtkinter
requests
pillow
